requirements.txt
fastapi[standard]>=0.109.0
langchain-core>=0.1.0
langchain-redis>=0.0.1
langchain-ollama>=0.0.1
langchain-groq>=0.0.1
langchain-community>=0.0.1
redis>=5.0.0
aio-pika>=9.3.0
python-dotenv>=1.0.0
boto3>=1.34.0
PyJWT>=2.8.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
PyPDF2>=3.0.0
botocore>=1.34.0
requests>=2.31.0
